EventCreationScreen moved to the correct package
